move `App` out of `components`
